feat: real-time signals (SIGRTMIN-SIGRTMAX) + siginfo_t SIGCHLD
- Expanded signal range to 64 (SIGRTMIN=32, SIGRTMAX=64)
- pending_signals and sig_mask changed from uint32_t to uint64_t
- PROCESS_SIG_MAX expanded to 65
- Added per-process sig_info[] array for siginfo_t delivery
- signal_send_info() now stores siginfo for signal_check retrieval
- SIGCHLD in process_exit_code sends proper siginfo_t (CLD_EXITED, pid, uid, status)
- Default action for real-time signals is ignore (POSIX)
- signal_mask/signal_unmask APIs updated to uint64_t
- syscall.c sigprocmask/sigpending updated for 64-bit masks
- process.h includes signal.h for struct siginfo definition

+++ src/test/test.c
@@ -511,10 +511,10 @@ static void test_signal(void) {
     t_ok("signal_register handler");
 
     /* Signal mask: mask SIGTERM, then send it — process should not terminate */
-    uint32_t old_mask = cur->sig_mask;
-    signal_mask((1u << SIGTERM));
+    uint64_t old_mask = cur->sig_mask;
+    signal_mask((1ULL << SIGTERM));
     ASSERT("signal_send masked SIGTERM OK", signal_send(cur->pid, SIGTERM) == 0);
-    signal_unmask((1u << SIGTERM));
+    signal_unmask((1ULL << SIGTERM));
     cur->sig_mask = old_mask;
     t_ok("signal mask/unmask");
 }
